+++ assignment-8/go.mod
@@ -1,0 +1,3 @@
+module github.com/adityapadekar-josh/assignment-8
+
+go 1.23.4
