+++ assignment-8/internal/api/domain.go
@@ -1,0 +1,5 @@
+package api
+
+type AddWebsiteToWatchListRequestBody struct {
+	Data []string `json:"data"`
+}
