+++ assignment-8/internal/api/router.go
@@ -1,0 +1,16 @@
+package api
+
+import (
+	"net/http"
+
+	"github.com/adityapadekar-josh/assignment-8/internal/models"
+)
+
+func NewRouter(dataStore *models.DataStore) *http.ServeMux {
+	router := http.NewServeMux()
+
+	router.HandleFunc("POST /websites", AddWebsiteToWatchList(dataStore))
+	router.HandleFunc("GET /websites", GetWebsitesStatus(dataStore))
+
+	return router
+}
