+++ assignment-8/internal/models/data_store.go
@@ -1,0 +1,8 @@
+package models
+
+import "sync"
+
+type DataStore struct {
+	Data map[string]string `json:"data"`
+	sync.RWMutex
+}
