+++ assignment-8/internal/pkg/response/domain.go
@@ -1,0 +1,6 @@
+package response
+
+type Response struct {
+	Message string      `json:"message"`
+	Result  interface{} `json:"result,omitempty"`
+}
