+++ assignment-8/internal/repository/init.go
@@ -1,0 +1,9 @@
+package repository
+
+import "github.com/adityapadekar-josh/assignment-8/internal/models"
+
+func InitDataStore() *models.DataStore {
+	return &models.DataStore{
+		Data: make(map[string]string),
+	}
+}
